requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.7.0
pandas>=2.0.0
pydantic>=2.5.0
PyYAML>=6.0.1
requests>=2.31.0

# PCB Processing
gerbonara==0.8.0
gerber-tools>=0.1.2
shapely>=2.0.0

# API & Web
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart==0.0.6

# LLM and RAG
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
openai>=1.12.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0
chromadb>=0.4.22

# Hugging Face integration (free alternative to OpenAI)
langchain-community>=0.0.13  # For HuggingFaceHub
langchain-huggingface>=0.0.1  # Updated Hugging Face integration
huggingface_hub>=0.19.0
transformers>=4.34.0  # For local model support
sentence-transformers>=2.2.0  # For embeddings

# Computer Vision & ML
opencv-python>=4.5.0
tensorflow>=2.8.0
torch>=1.9.0
torchvision>=0.10.0

# Feature Recognition
scikit-learn>=0.24.0
networkx>=2.6.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Utilities
tqdm>=4.62.0
python-dotenv>=1.0.0
json5>=0.9.14  # For enhanced JSON support
loguru>=0.7.0  # Better logging

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0 